+++ internal/models/models_test.go
@@ -1,4 +1,4 @@
-package main
+package models
 
 import "testing"
