+++ parser.go
@@ -4,16 +4,17 @@ import (
 	"io/ioutil"
 	"os"
 
+	"github.com/kampanosg/bakery/internal/models"
 	"gopkg.in/yaml.v2"
 )
 
-func ParseBakefile(f *os.File) (*Bakery, error) {
+func ParseBakefile(f *os.File) (*models.Bakery, error) {
 	content, err := ioutil.ReadAll(f)
 	if err != nil {
 		return nil, err
 	}
 
-	var r Bakery
+	var r models.Bakery
 	if err := yaml.Unmarshal(content, &r); err != nil {
 		return nil, err
 	}
